trading_backend/pyproject.toml
[tool.poetry]
name = "crypto-trading-system"
version = "0.1.0"
description = "Real-time cryptocurrency trading monitoring and prediction system"
authors = ["Devin <devin@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.104.1"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
python-binance = "^1.0.19"
xgboost = "^2.1.3"
pandas-ta = "^0.3.14b0"
jieba = "^0.42.1"
aiohttp = "^3.9.3"
pytesseract = "^0.3.13"
pillow = "^11.0.0"
opencv-python-headless = "^4.10.0.84"
python-multipart = "^0.0.20"
aiosmtplib = "^3.0.1"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
setuptools = "^69.0.3"
httpx = "0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
testpaths = ["tests"]
